🛡️ fix(mailbox): enforce dispatch requeue invariants
